// save-mate/src/app/core/enums/question-steps.enum.ts
export enum QuestionSteps {
    Income = "Income",
    SavedAmount  = "SavedAmount",
    Goals = "Goals",
    GoalDeadline = "GoalDeadline",
    RecurringExpenses = "RecurringExpenses",
    Expenses = "Expenses",
    PreviousMonthExpenses = "PreviousMonthExpenses",
    DebtQuestion = "DebtQuestion",
    Debts = "Debts",
    DebtDelay = "DebtDelay",
    Prepayment = "Prepayment"
}