// save-mate/src/app/core/enums/spending-categories-name.enum.ts
export enum SpendingCategoriesName {
    HousingAndUtilities = "Lakhatás és rezsi",
    Transportation = "Közlekedés",
    FoodAndHousehold = "Élelmiszer és háztartás",
    HealthAndPersonalExpenses = "Egészség és személyes kiadások",
    ClothingAndOtherShopping = "Ruházkodás és egyéb vásárlás",
    EntertainmentAndLeisure = "Szórakozás és szabadidő",
    ChildrenPerFamily = "Gyerekek/Család",
    LoansAndDebts = "Hitelek és adósságok",
    InsuranceAndFinancialProducts = "Biztosítások és pénzügyi termékek",
    SavingAndInvesting = "Megtakarítás és befektetés",
    OtherPerVariableExpenses = "Egyéb/Változó kiadások"
}